requirements.txt
mysql-connector-python==8.0.22
mysqlclient==2.0.1
protobuf==3.14.0
python-dotenv==0.15.0
six==1.15.0
